@@ -1,47 +1,35 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build"
     },
     "devDependencies": {
-        "@babel/preset-react": "^7.16.7",
         "autoprefixer": "^10.3.7",
         "axios": "^0.21",
-        "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
         "postcss": "^8.3.9",
-        "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
-        "sass-loader": "^11.0.1",
-        "tailwindcss": "^2.2.16",
-        "vue-loader": "15.9.8",
-        "vue-template-compiler": "^2.6.12"
+        "tailwindcss": "^2.2.16"
     },
     "dependencies": {
         "@tailwindcss/forms": "^0.3.4",
         "@tailwindcss/typography": "^0.4.1",
         "@toast-ui/vue-editor": "^3.2.0",
+        "@vitejs/plugin-vue": "^3.2.0",
         "boxicons": "^2.1.1",
         "chart.js": "2.9.4",
         "express": "^4.18.1",
         "imask": "^6.2.2",
+        "laravel-vite-plugin": "^0.7.0",
         "markdown-it": "^13.0.1",
         "material-file-icons": "^2.2.0",
         "prettier": "^2.6.2",
-        "react": "^18.1.0",
-        "react-dom": "^18.1.0",
         "shiki": "^0.11.1",
         "tippy.js": "^6.3.2",
-        "vue": "^2.6.14",
+        "vite": "^3.2.1",
+        "vue": "^3.2.41",
         "vue-chartjs": "^3.5.1",
-        "vue-tailwind": "^2.5.0",
-        "vue-tippy": "^4.12.0",
         "vuedraggable": "^2.24.3"
     }
 }
